requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydub>=0.25.1
gtts>=2.4.0
flask>=3.0.0
flask-cors>=4.0.0
librosa>=0.10.0
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.11.0
python-socketio>=5.9.0
python-engineio>=4.7.0
Pillow>=10.0.0